requirements.txt
Flask
python-dotenv
chardet
Frozen-Flask